Restructure eligibility form
