+++ packages/eligibility/src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
-import Eligibility from './Eligibility';
+import Eligibility from './components/Eligibility';
 import 'bootstrap/dist/css/bootstrap.min.css';
 
 ReactDOM.render(
